Refactor HomeView and add SpanishTexts implementation
- Rename HomeView.swift and update references in project.pbxproj
- Remove iOS-specific control size code from HomeView.swift
- Create new SpanishTexts.swift file implementing VoticeTextsProtocol
- Add HomeView.swift files for macOS and tvOS platforms
- Update project.pbxproj to include new HomeView and SpanishTexts files
- Update app icon images in Assets.xcassets with new versions

no message
