+++ Example/VoticeDemo/VoticeDemo_tvOS/HomeView.swift
@@ -1,0 +1,23 @@
+//
+//  HomeView.swift
+//  VoticeDemo_tvOS
+//
+//  Created by Arturo Carretero Calvo on 9/7/25.
+//  Copyright © 2025 ArtCC. All rights reserved.
+//
+
+import SwiftUI
+
+struct HomeView: View {
+    // MARK: - Properties
+
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+// MARK: - Preview
+
+#Preview {
+    HomeView()
+}
